+++ apps/web/src/app/layout.tsx
@@ -1,12 +1,13 @@
-import type { Metadata } from "next";
-import { Toaster } from "@/components/ui/sonner"
-import "./globals.css";
-import {ThemeProvider} from "next-themes";
-import AuthProvider from "@/lib/context/auth-context";
+import type { Metadata } from 'next';
+import { Toaster } from '@/components/ui/sonner';
+import './globals.css';
+import { ThemeProvider } from 'next-themes';
+import AuthProvider from '@/lib/context/auth-context';
+import { QueryProvider } from '@/lib/providers/query-provider';
 
 export const metadata: Metadata = {
-  title: "Rawstack | Public Website",
-  description: "The client facing website starter",
+  title: 'Rawstack | Public Website',
+  description: 'The client facing website starter',
 };
 
 export default function RootLayout({
@@ -17,17 +18,12 @@ export default function RootLayout({
   return (
     <html lang="en" suppressHydrationWarning>
       <body className="antialiased">
-      <ThemeProvider
-          attribute="class"
-          defaultTheme="system"
-          enableSystem
-          disableTransitionOnChange
-      >
-          <AuthProvider>
-            {children}
-          </AuthProvider>
-      </ThemeProvider>
-      <Toaster />
+        <QueryProvider>
+          <ThemeProvider attribute="class" defaultTheme="system" enableSystem disableTransitionOnChange>
+            <AuthProvider>{children}</AuthProvider>
+          </ThemeProvider>
+        </QueryProvider>
+        <Toaster />
       </body>
     </html>
   );